# docs/swagger.yaml
basePath: /
definitions:
...
  dto.ItemResponse:
    properties:
      code:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
...
  dto.ModuleResponse:
    properties:
      id:
        type: integer
      name:
        type: string
      slug:
        type: string
    type: object
...
  dto.ProfileModuleItemAttachRequest:
    properties:
      item_id:
        example: 1
        type: integer
    required:
    - item_id
    type: object
  dto.ProfileModuleItemSyncRequest:
    properties:
      items:
        items:
          type: integer
        type: array
    type: object
  dto.ProfileModuleSyncRequest:
    properties:
      modules:
        items:
          type: integer
        type: array
    type: object
...
host: localhost:8082
info:
  contact: {}
  description: API Backend para GOLANG DEMO.
  title: Golang DEMO
  version: "1.0"
paths:
...
  /profiles/{id}/modules:
    get:
      description: Retorna los módulos asignados a un perfil específico
      parameters:
      - description: ID del perfil
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.ModuleResponse'
            type: array
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Listar módulos asignados a un perfil
      tags:
      - Profile Modules
    put:
      consumes:
      - application/json
      description: Asigna o remueve módulos de un perfil (operación sync)
      parameters:
      - description: ID del perfil
        in: path
        name: id
        required: true
        type: integer
      - description: Array de IDs de módulos
        in: body
        name: modules
        required: true
        schema:
          $ref: '#/definitions/dto.ProfileModuleSyncRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Sincronizar módulos de un perfil
      tags:
      - Profile Modules
  /profiles/{id}/modules/{moduleId}/items:
    get:
      description: Retorna los items/permisos asignados a un módulo específico de
        un perfil
      parameters:
      - description: ID del perfil
        in: path
        name: id
        required: true
        type: integer
      - description: ID del módulo
        in: path
        name: moduleId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.ItemResponse'
            type: array
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Listar items de un módulo de perfil
      tags:
      - Profile Module Items
    post:
      consumes:
      - application/json
      description: Asigna un item/permiso específico sin afectar los demás
      parameters:
      - description: ID del perfil
        in: path
        name: id
        required: true
        type: integer
      - description: ID del módulo
        in: path
        name: moduleId
        required: true
        type: integer
      - description: ID del item
        in: body
        name: item
        required: true
        schema:
          $ref: '#/definitions/dto.ProfileModuleItemAttachRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.ItemResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Agregar item individual a un módulo de perfil
      tags:
      - Profile Module Items
    put:
      consumes:
      - application/json
      description: Asigna o remueve items/permisos de un módulo específico (sync)
      parameters:
      - description: ID del perfil
        in: path
        name: id
        required: true
        type: integer
      - description: ID del módulo
        in: path
        name: moduleId
        required: true
        type: integer
      - description: Array de IDs de items
        in: body
        name: items
        required: true
        schema:
          $ref: '#/definitions/dto.ProfileModuleItemSyncRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Sincronizar items de un módulo de perfil
      tags:
      - Profile Module Items
  /profiles/{id}/modules/{moduleId}/items/{itemId}:
    delete:
      description: Remueve un item/permiso específico de un módulo de perfil
      parameters:
      - description: ID del perfil
        in: path
        name: id
        required: true
        type: integer
      - description: ID del módulo
        in: path
        name: moduleId
        required: true
        type: integer
      - description: ID del item
        in: path
        name: itemId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Eliminar item de un módulo de perfil
      tags:
      - Profile Module Items
...
schemes:
- http
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"